updated some api routes

+++ src/main/java/com/vijaybrothers/store/repository/PaymentRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
 public interface PaymentRepository extends JpaRepository<Payment,Integer> {
-    Optional<Payment> findByOrderOrderId(Integer orderId);
+    Optional<Payment> findByOrderOrderId(Long orderId);
     Optional<Payment> findByTransactionId(String transactionId);
 }
